Halved pitch of tick tone. This increases the signal to noise ratio because speakers are able to produce this lower frequency at a much higher amplitude. This also results in more samples per tick, which allows the auto threshold to target the first edge instead of the largest edge, which works better for echoes.

+++ Shared-Items/RecordingAnalyzer.h
@@ -33,7 +33,7 @@ public:
 
 	static RecordingResult AnalyzeRecording(const GeneratedSamples& generatedSamples, const WasapiInput& input, AudioFormat* format, OutputOffsetProfile* currentProfile, DacLatencyProfile* referenceDacLatency);
 	static std::vector<AveragedResult> AnalyzeResults(std::vector<RecordingResult> results, time_t tTime, const AudioEndpoint& outputEndpoint);
-	static std::vector<TickPosition> GetTicks(float* recordedSamples, int recordedSamplesLength, int sampleRate, int expectedTickFrequency, int numTicks, float threshold);
+	static std::vector<TickPosition> GetTicks(float* recordedSamples, int recordedSamplesLength, int sampleRate, double expectedTickFrequency, int numTicks, float threshold);
 
 	static void SaveRecording(const WasapiInput& input, std::string path, std::string filename);
 	static void SaveIndividualResult(IResultsWriter& writer, const AudioEndpoint& outputEndpoint, const AudioEndpoint& inputEndpoint, RecordingResult& result, std::string testRootPath, std::string inputFormat);
